% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_epochs_out.R
\name{read_epochs_out}
\alias{read_epochs_out}
\title{Read in REAPER epochs output file as tibble}
\usage{
read_epochs_out(epochsFile, audioFile = NULL)
}
\arguments{
\item{epochsFile}{String giving the file name of an ASCII epochs / pitchmarks
file created by REAPER.}

\item{audioFile}{String giving the name of the original audio file. If a
string is provided, the function will return a named list instead of a
vector (practical if creating a large object with many of these files).
Default is `NULL`.}
}
\value{
# A numeric vector of epochs, or a named list with this vector.
}
\description{
Reads in an ASCII epochs / pitchmarks file created by REAPER as a
vector of epochs
}
\examples{
path <- file.path(system.file(package = 'reapeR'), 'extdata', 'epochs')
epochs <- read_epochs_out(path)
}
