% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapeR-package.R
\docType{package}
\name{reapeR-package}
\alias{reapeR}
\alias{reapeR-package}
\title{reapeR: Robust Epoch and Pitch Estimation in R}
\description{
Wrapper for David Talkin's 'Robust Epoch and Pitch EstimatoR' (REAPER) software for estimating pitch and glottal closure instants from sound files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rpuggaardrode/reapeR}
  \item Report bugs at \url{https://github.com/rpuggaardrode/reapeR/issues}
}

}
\author{
\strong{Maintainer}: Rasmus Puggaard-Rode \email{rasmus.puggaard-rode@ling-phil.ox.ac.uk} [copyright holder]

}
\keyword{internal}
