% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaper.R
\name{reaper}
\alias{reaper}
\title{Estimate pitch and/or epochs with REAPER}
\usage{
reaper(
  file,
  start = 0,
  end = Inf,
  channel = 1,
  f0min = 40,
  f0max = 500,
  interval = 0.005,
  hilbert = FALSE,
  suppress_highpass_filter = FALSE,
  unvoiced_cost = 0.9,
  output = c("pitch", "epochs"),
  verbose = TRUE
)
}
\arguments{
\item{file}{String giving the name of a WAV file to analyze}

\item{start}{Numeric value giving the analysis start time in seconds.
When possible, the actual analysis will begin 50 ms prior to this time.
Default is `0`.}

\item{end}{Numeric value giving the analysis end time in seconds.
When possible, the actual analysis will begin 50 ms after this time.
Default is `Inf`.}

\item{channel}{Numeric value specifying the channel to be analyzed.
Default is `1`.}

\item{f0min}{Numeric value specifying pitch floor. Default is `40`.}

\item{f0max}{Numeric value specifying pitch ceiling. Default is `500`.}

\item{interval}{Numeric value giving the F0 output interval in seconds.
Default is `0.005`.}

\item{hilbert}{Boolean; should Hilbert transform be applied prior to
analysis? Default is `FALSE`.}

\item{suppress_highpass_filter}{Boolean; should highpass filter be
suppressed? Default is `FALSE`.}

\item{unvoiced_cost}{Numeric; cost for unvoiced segments. Default is `0.9`,
set higher value to estimate more F0 in noise.}

\item{output}{String or vector of strings specifying which estimates to
output. Possible values are `pitch` and `epochs`, default is to output
both.}

\item{verbose}{Boolean; should messages be printed to the console?
Default is `TRUE`.}
}
\value{
If `output = 'pitch'`, returns a data frame with four columns:
* `time`, giving the frame time in seconds
* `voiced` Boolean specifying whether that frame is `voiced`
* `f0` F0 in Hz
* `file` String giving the name of the analyzed file.

If `output = 'epochs'`, returns a numeric vector of epochs, or glottal
closure instants.

If `output = c('pitch', 'epochs')`, returns a list with the above two
outputs.
}
\description{
Track pitch and/or estimate epochs using David Talkin's REAPER library.
}
\examples{
snd <- file.path(system.file('extdata', package = 'reapeR'), '1.wav')
vals <- reaper(snd)
}
