% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaper2ssff.R
\name{reaper2ssff}
\alias{reaper2ssff}
\title{Convert REAPER output to SSFF format}
\usage{
reaper2ssff(reaper_output)
}
\arguments{
\item{reaper_output}{Output from [reaper] containing pitch measures.}
}
\value{
An object of class `AsspDataObj`.
}
\description{
Convert REAPER output for a single file to Simple Signal File Format data.
Usually used by [reaper2emuDB], but may also be used as a stand-alone for
e.g. plotting REAPER pitch data with `praatpicture`.
}
\examples{
snd <- file.path(system.file('extdata', package = 'reapeR'), '1.wav')
vals <- reaper(snd)
ssffObj <- reaper2ssff(vals)
}
