% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaper_bulk.R
\name{reaper_bulk}
\alias{reaper_bulk}
\title{Bulk estimate pitch and/or epochs with with REAPER}
\usage{
reaper_bulk(
  directory,
  output = c("pitch", "epochs"),
  f0min = 40,
  f0max = 500,
  hirst2pass = FALSE,
  praat_output = FALSE,
  praat_output_dir = NULL,
  hirst2pass_f0min = 60,
  hirst2pass_f0max = 700,
  ...
)
}
\arguments{
\item{directory}{String giving the name of a directory where all WAV files
should be processed. Alternatively the handle of a loaded EMU database.}

\item{output}{String or vector of strings specifying which estimates to
output. Default is to output `pitch` and `epochs`. Possible values are:
* `pitch` Returning a data frame with equidistant pitch estimates at
intervals determined by `interval`.
* `epochs` returning a named list with a numeric vector specifying the times
of estimated glottal closure instants.
* `gci_cand` Returning a data frame with times that are candidate glottal
closure instants, as well as their time, local pitch value if determined
to be voiced, and normalized cross-correlation functions (`nccf`).
* `probs` Returning a data frame with pseudo-probabilities of voicing,
voicing onset, and voicing offset at regular intervals, as well as bandpassed
(low-frequency) root-mean-squared amplitude of the signal, which is used to
derive the pseudo-probability of voicing.}

\item{f0min}{Numeric value specifying pitch floor. Default is `40`.}

\item{f0max}{Numeric value specifying pitch ceiling. Default is `500`.}

\item{hirst2pass}{Boolean; should pitch floor and ceiling be dynamically
estimated using the two-pass procedure proposed by Hirst and de Looze?
In this case, REAPER is first run with liberal floor and ceiling values of
60 Hz and 700 Hz respectively, and then rerun with optimal floor and ceiling
values estimated from the first and third quantiles of the first pass.
(Floor = 0.75 Q1, ceiling = 1.5 Q3). If `directory` refers to an EMU database
with multiple sessions, the two-pass procedure is run separately for each
session.}

\item{praat_output}{Boolean; should REAPER output be stored as a Praat
`.Pitch` file? Default is `FALSE`.}

\item{praat_output_dir}{String giving the location of a directory where
Praat `.Pitch` files should be stored. Default is `NULL`.}

\item{hirst2pass_f0min}{Numeric giving the pitch floor in Hz for the first
pass in a two-pass procedure. Default is `60`.}

\item{hirst2pass_f0max}{Numeric giving the pitch ceiling in Hz for the first
pass in a two-pass procedure. Default is `700`.}

\item{...}{Further arguments passed on to `reaper()`.}
}
\value{
Depending on the value of `output`, either a nested list of outputs,
a data frame, or a numeric vector.
}
\description{
Wrapper function to call David Talkin's REAPER software on all WAV files
in a directory
}
\examples{
dir <- file.path(system.file('extdata', package = 'reapeR'))
vals <- reaper_bulk(dir)
}
