% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaper_install.R
\name{reaper_install}
\alias{reaper_install}
\title{Attempt installation of REAPER}
\usage{
reaper_install()
}
\value{
Used for side effects.
}
\description{
This function can be run if the installation of REAPER failed when
installing the package.
}
\examples{
\dontrun{
reaper_install()
}
}
