% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reaper_wrap}
\alias{reaper_wrap}
\title{Wrap REAPER C++ library}
\usage{
reaper_wrap(
  samples,
  sample_rate,
  f0min = 40,
  f0max = 500,
  suppress_highpass_filter = FALSE,
  hilbert = FALSE,
  interval = 0.005,
  unvoiced_cost = 0.9,
  unvoiced_pulse_interval = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{samples}{Numeric vector giving audio samples in 16-bit integer format.}

\item{sample_rate}{Numeric giving the sampling rate of `samples`.}

\item{f0min}{Numeric value specifying pitch floor. Default is `40`.}

\item{f0max}{Numeric value specifying pitch ceiling. Default is `500`.}

\item{suppress_highpass_filter}{Boolean; should highpass filter be
suppressed? Default is `FALSE`.}

\item{hilbert}{Boolean; should Hilbert transform be applied prior to
analysis? Default is `FALSE`.}

\item{interval}{Numeric value giving the F0 output interval in seconds.
Default is `0.005`.}

\item{unvoiced_cost}{Numeric; cost for unvoiced segments. Default is `0.9`,
set higher value to estimate more F0 in noise.}

\item{unvoiced_pulse_interval}{Numeric; what should be the interval in
seconds of epoch pulses outside of voiced intervals? Default is `0.01`.}

\item{verbose}{Boolean; should messages be printed to the console?}
}
\value{
A named list containing all of the output from the
`epoch_tracker` procedure.
}
\description{
Low-level wrapper for calling REAPER's `epoch_tracker` procedure
}
\examples{
file <- file.path(system.file('extdata', package = 'reapeR'), '1.wav')
snd <- tuneR::readWave(file)
results <- reaper_wrap(snd@left, snd@samp.rate)

}
