% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_praat_epochs.R
\name{write_praat_epochs}
\alias{write_praat_epochs}
\title{Write REAPER output to Praat PointProcess file}
\usage{
write_praat_epochs(reaper_out, directory, filename)
}
\arguments{
\item{reaper_out}{Data frame with the output of REAPER, as loaded with
[read_epochs_out] or created with [reaper].}

\item{directory}{String giving the location to use for storing the resulting
`.PointProcess` file.}

\item{filename}{String giving the file name of the resulting file (note that
the `.PointProcess` extension is added automatically).}
}
\value{
Used for side effects.
}
\description{
Convert the estimated epoch locations of REAPER to a `.PointProcess` file
that can be read by Praat.
}
\examples{
\dontrun{
snd <- file.path(system.file('extdata', package = 'reapeR'), '1.wav')
epo <- reaper(snd, output = 'epochs')
out_path <- file.path(tempdir(), 'epo.PointProcess')
file.exists(out_path)
write_praat_epochs(epo, tempdir(), 'epo')
file.exists(out_path)
}
}
