% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_praat_pitch.R
\name{write_praat_pitch}
\alias{write_praat_pitch}
\title{Write REAPER output to Praat Pitch file}
\usage{
write_praat_pitch(reaper_out, directory, filename)
}
\arguments{
\item{reaper_out}{Data frame with the output of REAPER, as loaded with
[read_pitch_out] or created with [reaper].}

\item{directory}{String giving the location to use for storing the resulting
`.Pitch` file.}

\item{filename}{String giving the file name of the resulting file (note that
the `.Pitch` extension is added automatically).}
}
\value{
Used for side effects.
}
\description{
Convert the pitch estimates of REAPER to a `.Pitch` file that can be read by
Praat.
}
\examples{
\dontrun{
snd <- file.path(system.file('extdata', package = 'reapeR'), '1.wav')
pit <- reaper(snd, output = 'pitch')
out_path <- file.path(tempdir(), 'pit.Pitch')
file.exists(out_path)
write_praat_pitch(pit, tempdir(), 'pit')
file.exists(out_path)
}
}
