% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pitch_out.R
\name{read_pitch_out}
\alias{read_pitch_out}
\title{Read in REAPER pitch output file as tibble}
\usage{
read_pitch_out(pitchFile, audioFile = NULL, delete = FALSE)
}
\arguments{
\item{pitchFile}{String giving the file name of an ASCII pitch file
created by REAPER.}

\item{audioFile}{String giving the name of the original audio file. If a
string is provided, a separate column will be made in the tibble repeating
this string (practical if creating a large data frame with many of these
files). Default is `NULL`, in which case no such column is created.}

\item{delete}{Boolean; should the original REAPER output file be deleted?
Default is `FALSE`.}
}
\value{
A data frame with four columns:
* `time`, giving the frame time in seconds
* `voiced` Boolean specifying whether that frame is `voiced`
* `f0` F0 in Hz
* `file` String giving the name of the analyzed file.
}
\description{
Reads in an ASCII pitch file created by REAPER as a well-formatted tibble
}
\examples{
path <- file.path(system.file(package = 'reapeR'), 'extdata', 'pitch')
pitch <- read_pitch_out(path)
}
