% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaper.R
\name{reaper}
\alias{reaper}
\title{Estimate pitch and/or epochs with REAPER}
\usage{
reaper(
  file,
  start = 0,
  end = Inf,
  channel = 1,
  f0min = 40,
  f0max = 500,
  interval = 0.005,
  hilbert = FALSE,
  suppress_highpass_filter = FALSE,
  unvoiced_cost = 0.9,
  output = c("pitch", "epochs"),
  force_list_output = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{file}{String giving the name of a WAV file to analyze}

\item{start}{Numeric value giving the analysis start time in seconds.
When possible, the actual analysis will begin 50 ms prior to this time.
Default is `0`.}

\item{end}{Numeric value giving the analysis end time in seconds.
When possible, the actual analysis will begin 50 ms after this time.
Default is `Inf`.}

\item{channel}{Numeric value specifying the channel to be analyzed.
Default is `1`.}

\item{f0min}{Numeric value specifying pitch floor. Default is `40`.}

\item{f0max}{Numeric value specifying pitch ceiling. Default is `500`.}

\item{interval}{Numeric value giving the F0 output interval in seconds.
Default is `0.005`.}

\item{hilbert}{Boolean; should Hilbert transform be applied prior to
analysis? Default is `FALSE`.}

\item{suppress_highpass_filter}{Boolean; should highpass filter be
suppressed? Default is `FALSE`.}

\item{unvoiced_cost}{Numeric; cost for unvoiced segments. Default is `0.9`,
set higher value to estimate more F0 in noise.}

\item{output}{String or vector of strings specifying which estimates to
output. Default is to output `pitch` and `epochs`. Possible values are:
* `pitch` Returning a data frame with equidistant pitch estimates at
intervals determined by `interval`.
* `epochs` returning a named list with a numeric vector specifying the times
of estimated glottal closure instants.
* `gci_cand` Returning a data frame with times that are candidate glottal
closure instants, as well as their time, local pitch value if determined
to be voiced, and normalized cross-correlation functions (`nccf`).
* `probs` Returning a data frame with pseudo-probabilities of voicing,
voicing onset, and voicing offset at regular intervals, as well as bandpassed
(low-frequency) root-mean-squared amplitude of the signal, which is used to
derive the pseudo-probability of voicing.
* `resids` Returning a data frame with the original signal, linear prediction
residuals, and linear prediction residuals normalized by local
amplitude.}

\item{force_list_output}{Boolean; if `TRUE`, the function returns a named
nested list even if only one output is chosen. Default is `FALSE`.}

\item{verbose}{Boolean; should diagnostic messages be printed to the console?
Default is `FALSE`.}
}
\value{
Depending on the value of `output`, either a nested list of outputs,
a data frame, or a numeric vector.
}
\description{
Track pitch and/or estimate epochs using David Talkin's REAPER library.
}
\examples{
snd <- file.path(system.file('extdata', package = 'reapeR'), '1.wav')
vals <- reaper(snd)
}
