% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaper.R
\name{reaper}
\alias{reaper}
\title{Estimate pitch and/or epochs with REAPER}
\usage{
reaper(
  file,
  f0min = 40,
  f0max = 500,
  interval = 0.005,
  hilbert = FALSE,
  suppress_highpass_filter = FALSE,
  unvoiced_cost = 0.9,
  output = c("pitch", "epochs")
)
}
\arguments{
\item{file}{String giving the name of an audio file to analyze}

\item{f0min}{Numeric value specifying pitch floor. Default is `40`.}

\item{f0max}{Numeric value specifying pitch ceiling. Default is `500`.}

\item{interval}{Numeric value giving the F0 output interval in seconds.
Default is `0.005`.}

\item{hilbert}{Boolean; should Hilbert transform be applied prior to
analysis? Default is `FALSE`.}

\item{suppress_highpass_filter}{Boolean; should highpass filter be
suppressed? Default is `FALSE`.}

\item{unvoiced_cost}{Numeric; cost for unvoiced segments. Default is `0.9`,
set higher value to estimate more F0 in noise.}

\item{output}{String or vector of strings specifying which estimates to
output. Possible values are `pitch` and `epochs`, default is to output
both.}
}
\value{
If `output = 'pitch'`, returns a data frame with four columns:
* `time`, giving the frame time in seconds
* `voiced` Boolean specifying whether that frame is `voiced`
* `f0` F0 in Hz
* `file` String giving the name of the analyzed file.

If `output = 'epochs'`, returns a numeric vector of epochs, or glottal
closure instants.

If `output = c('pitch', 'epochs')`, returns a list with the above two
outputs.
}
\description{
Wrapper function to call David Talkin's REAPER software in R
}
\examples{
snd <- file.path(system.file('extdata', package = 'reapeR'), '1.wav')
vals <- reaper(snd)
}
