% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaper2emuDB.R
\name{reaper2emuDB}
\alias{reaper2emuDB}
\title{Import REAPER output to an EMU database as SSFF files}
\usage{
reaper2emuDB(reaper_output, db_handle)
}
\arguments{
\item{reaper_output}{Data frame or list object created with the
`reaper_bulk()` function run over a loaded EMU database.}

\item{db_handle}{Handle of a loaded EMU database}
}
\value{
Nothing; run for side effects.
}
\description{
When pitch has been tracked over an EMU database, this function is used to
import the measures to the database in the Simple Signal File Format.
}
\examples{
\dontrun{
emuR::create_emuRdemoData(tempdir())
db <- emuR::load_emuDB(file.path(tempdir(), 'emuR_demoData', 'ae_emuDB'))
emuR::list_ssffTrackDefinitions(db)
out <- reaper_bulk(db)
reaper2emuDB(out, db)
emuR::list_ssffTrackDefinitions(db)
}
}
