% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaper2ssff.R
\name{reaper2ssff}
\alias{reaper2ssff}
\title{Convert REAPER output to SSFF files and import to EMU database}
\usage{
reaper2ssff(reaper_output, db_handle)
}
\arguments{
\item{reaper_output}{Data frame or list object created with the
`reaper_bulk()` function.}

\item{db_handle}{Handle of a loaded EMU database}
}
\value{
Nothing; run for side effects.
}
\description{
Convert REAPER output to SSFF files and import to EMU database
}
\examples{
# not now
}
