% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaper_bulk.R
\name{reaper_bulk}
\alias{reaper_bulk}
\title{Bulk estimate pitch and/or epochs with with REAPER}
\usage{
reaper_bulk(
  directory,
  output = c("pitch", "epochs"),
  f0min = 40,
  f0max = 500,
  hirst2pass = FALSE,
  praat_output = FALSE,
  praat_output_dir = NULL,
  ...
)
}
\arguments{
\item{directory}{String giving the name of a directory where all WAV files
should be processed. Alternatively the handle of a loaded EMU database.}

\item{output}{String or vector of strings specifying which estimates to
output. Possible values are `pitch` and `epochs`, default is to output
both.}

\item{f0min}{Numeric value specifying pitch floor. Default is `40`.}

\item{f0max}{Numeric value specifying pitch ceiling. Default is `500`.}

\item{hirst2pass}{Boolean; should pitch floor and ceiling be dynamically
estimated using the two-pass procedure proposed by Hirst and de Looze?
In this case, REAPER is first run with liberal floor and ceiling values of
60 Hz and 700 Hz respectively, and then rerun with optimal floor and ceiling
values estimated from the first and third quantiles of the first pass.
(Floor = 0.75 Q1, ceiling = 1.5 Q3). If `directory` refers to an EMU database
with multiple sessions, the two-pass procedure is run separately for each
session.}

\item{praat_output}{Boolean; should REAPER output be stored as a Praat
`.Pitch` file? Default is `FALSE`.}

\item{praat_output_dir}{String giving the location of a directory where
Praat `.Pitch` files should be stored. Default is `NULL`.}

\item{...}{Further arguments passed on to `reaper()`.}
}
\value{
If `output = 'pitch'`, returns a data frame with four columns:
* `time`, giving the frame time in seconds
* `voiced` Boolean specifying whether that frame is `voiced`
* `f0` F0 in Hz
* `file` String giving the name of the analyzed file.

If `directory` refers to an EMU database, furthermore contains the columns
`session` and `bundle`.

If `output = 'epochs'`, returns a numeric vector of epochs, or glottal
closure instants.

If `output = c('pitch', 'epochs')`, returns a list with the above two
outputs.
}
\description{
Wrapper function to call David Talkin's REAPER software on all WAV files
in a directory
}
\examples{
dir <- file.path(system.file('extdata', package = 'reapeR'))
vals <- reaper(dir)
}
